+++ week2/assignment/가계부 💸/js/initDataRender.js
@@ -1,6 +1,6 @@
 import {
   renderHistoryList,
-} from "./utils/hisoryDataRender.js";
+} from "./utils/historyDataRender.js";
 
 import { 
   renderAssetData,
